Addons: fix a few pedantic warnings
git-svn-id: https://geany-plugins.svn.sourceforge.net/svnroot/geany-plugins/trunk/geany-plugins@2174 e8383189-8249-0410-b506-acc975200cbd

+++ addons/src/ao_xmltagging.c
@@ -88,7 +88,7 @@ void ao_xmltagging(void)
 			tag = g_strdup(gtk_entry_get_text(GTK_ENTRY(textbox)));
 			if (NZV(tag))
 			{
-				gint end = 0;
+				gsize end = 0;
 				GString *tmp = NULL;
 				const gchar *end_tag;
